requirements.txt
pyserial>=3.5
rich>=13.7.0
